+++ src/main/java/graphql/annotations/connection/exceptions/GraphQLConnectionException.java
@@ -12,11 +12,11 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  */
-package graphql.annotations.connection;
+package graphql.annotations.connection.exceptions;
 
-class GraphQLConnectionException extends RuntimeException {
+public class GraphQLConnectionException extends RuntimeException {
 
-    GraphQLConnectionException(String error) {
+    public GraphQLConnectionException(String error) {
         super(error);
     }
 }
